Fix functionality when resigning a game
When resigning a game, the modal wasn't properly updating who won.
So when a player resigns, we now check if the player is actually in the
game, and if so, we update the game to show the other player as the
winner.

Also, accidentally committed the test rows in the move history table,
so I removed those, oops.