Remove extra JavaScript from index.js
Before, we sent a POST request to /games/create using fetch, which
would then return a json response that we would use to redirect to the
user create page. Now, we remove this unnecessary JavaScript button and
just use a regular HTML form to send a POST request to /games/create,
and then redirect to the user create page if needed. Less hacky
JavaScript and less code overall, yippee!