Rework how error messages are displayed
Show error messages in a div that floats above the page content and
automatically disappears after a few seconds. Make it so that errors
that come from sockets are displayed in the same way as errors that come
from routes. Also move the socket from the main.js file to the base.js
file, so that anytime a socket is needed, it can be imported from the
base.js file that is always loaded on every page.