Fix issue with decorators
function.__code__.co_varnames didn't work fully for multiple decorators
because the co_varnames tuple would only have the other decorator's
arguments. So instead we replaced it with
inspect.signature(function).parameters to get access to the original function's
arguments and to properly conditionally pass kwargs from the decorator
to the original function.